BaseControllerを実装し、app/controllers/admin配下にあるコントローラはこのコントローラを継承するようにする
