+++ app/controllers/admin/admin_users_controller.rb
@@ -1,5 +1,4 @@
-class Admin::AdminUsersController < ApplicationController
-  before_action :check_login
+class Admin::AdminUsersController < Admin::BaseController
   before_action :prohibit_access_for_regular_admin, only: %i[ index show new create edit update destroy ]
   before_action :set_admin_user, only: %i[ show edit update destroy ]
