+++ app/controllers/admin/stores_controller.rb
@@ -1,5 +1,4 @@
-class Admin::StoresController < ApplicationController
-  before_action :check_login
+class Admin::StoresController < Admin::BaseController
   before_action :set_brand, only: %i[ new create show edit update destroy ]
   before_action :set_store, only: %i[ show edit update destroy ]
   before_action :prohibit_access_for_regular_admin, only: %i[ show new create edit destroy ]
